+++ crates/miden-testing/tests/scripts/send_note.rs
@@ -7,7 +7,7 @@ use miden_protocol::note::{
     Note,
     NoteAssets,
     NoteAttachment,
-    NoteAttachmentType,
+    NoteAttachmentScheme,
     NoteInputs,
     NoteMetadata,
     NoteRecipient,
@@ -38,7 +38,7 @@ async fn test_send_note_script_basic_wallet() -> anyhow::Result<()> {
 
     let tag = NoteTag::with_account_target(sender_basic_wallet_account.id());
     let elements = [9, 8, 7, 6, 5u32].map(Felt::from).to_vec();
-    let attachment = NoteAttachment::new_array(NoteAttachmentType::new(42), elements.clone())?;
+    let attachment = NoteAttachment::new_array(NoteAttachmentScheme::new(42), elements.clone())?;
     let metadata = NoteMetadata::new(sender_basic_wallet_account.id(), NoteType::Public, tag)
         .with_attachment(attachment.clone());
     let assets = NoteAssets::new(vec![sent_asset]).unwrap();
@@ -98,7 +98,7 @@ async fn test_send_note_script_basic_fungible_faucet() -> anyhow::Result<()> {
         AccountInterface::from_account(&sender_basic_fungible_faucet_account);
 
     let tag = NoteTag::with_account_target(sender_basic_fungible_faucet_account.id());
-    let attachment = NoteAttachment::new_word(NoteAttachmentType::new(100), Word::empty());
+    let attachment = NoteAttachment::new_word(NoteAttachmentScheme::new(100), Word::empty());
     let metadata =
         NoteMetadata::new(sender_basic_fungible_faucet_account.id(), NoteType::Public, tag)
             .with_attachment(attachment);